+++ Phoenix/PHApp.m
@@ -121,8 +121,7 @@
 
     NSPredicate *visibility = [NSPredicate predicateWithBlock:^BOOL (PHWindow *window,
                                                                      __unused NSDictionary<NSString *, id> *bindings) {
-
-        return ![[window app] isHidden] && [window isNormal] && ![window isMinimized];
+        return [window isVisible];
     }];
 
     return [[self windows] filteredArrayUsingPredicate:visibility];